code, decl
